refactor: modify components name
