--- src/components/common/GroupPlanResultList/index.ts
@@ -1,1 +1,0 @@
-export * from './GroupPlanResultList';
