--- src/components/common/SinglePlanResultList copy/index.ts
@@ -1,1 +1,0 @@
-export * from './SinglePlanResultList';
